Extract `buildSite` test method to TestCase
