+++ tests/TestCase.php
@@ -5,14 +5,16 @@ namespace Tests;
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use TightenCo\Jigsaw\File\Filesystem;
 use TightenCo\Jigsaw\File\InputFile;
+use TightenCo\Jigsaw\Jigsaw;
+use org\bovigo\vfs\vfsStream;
 
 class TestCase extends BaseTestCase
 {
     public $app;
     public $filesystem;
     public $tempPath;
-    protected $sourcePath = __DIR__ . '/source';
-    protected $destinationPath = __DIR__ . '/build_testing';
+    public $sourcePath = __DIR__ . '/source';
+    public $destinationPath = __DIR__ . '/build_testing';
 
     public function setUp()
     {
@@ -34,14 +36,14 @@ class TestCase extends BaseTestCase
         parent::tearDown();
     }
 
-    protected function prepareTempDirectory()
+    public function prepareTempDirectory()
     {
         if (! $this->filesystem->isDirectory($this->tempPath)) {
             $this->filesystem->makeDirectory($this->tempPath, 0755, true);
         }
     }
 
-    protected function cleanupTempDirectory()
+    public function cleanupTempDirectory()
     {
         $this->filesystem->deleteDirectory($this->tempPath);
     }
@@ -51,5 +53,24 @@ class TestCase extends BaseTestCase
         $sourceFile = $this->filesystem->getFile(str_finish($this->sourcePath, '/') . pathinfo($filename)['dirname'], basename($filename));
 
         return new InputFile($sourceFile, $this->sourcePath);
+    }
+
+    public function setupSource($source = [])
+    {
+        return vfsStream::setup('virtual', null, ['source' => $source]);
+    }
+
+    public function buildSite($vfs, $config = [])
+    {
+        $this->app->config = collect($config);
+        $this->app->buildPath = [
+            'source' => $vfs->url() . '/source',
+            'destination' => $vfs->url() . '/build',
+        ];
+
+        $jigsaw = $this->app->make(Jigsaw::class);
+        $jigsaw->build('test');
+
+        return $jigsaw;
     }
 }
